refactor(voyage): les manifestes de jambe sortent d'app.js
`voyage-donnees.ts` porte la moitié DONNÉE du bloc : calculer le manifeste d'une jambe
(`legManifest`, `legFeeCtx`), lire et persister l'intention de l'utilisateur (`legIntent`,
`legEffectiveLines`, les deux stores et leurs clés). `findCommodity` rejoint `marche.ts`, dont il
est une dérivation.

Les ACTIONS restent dans `app.js` — déplier un éditeur, changer des SCU, ajouter une suggestion.
Elles enchaînent une mutation ET un rendu, et c'est le rendu qui les y retient ; elles partiront
quand la vue Voyage emménagera dans l'arbre. Même partage que pour les corrections, la persistance
et le chargement.

UNE RÉFÉRENCE VIVE SORT DE CE MODULE, ET C'EST ÉCRIT EN TÊTE. `legIntent` rend
`etat.JOURNEY_EDITS[k]` TEL QUEL, et l'appelant le mute. C'est l'une des quatre raisons pour
lesquelles `etat.ts` ne notifie pas à l'écriture — aucun accesseur ne verrait cette écriture. Rendre
une copie « pour faire propre » casserait les mutants, qui comptent sur l'identité et appellent
`saveJourneyEdits()` eux-mêmes.

C'est la clé de voûte qui bloquait trois vues : le Plan de vol (`planData` appelle
`legEffectiveLines`), les Commodités (`journeyCarriedCommodities`) et le compagnon.

`app.js` : 3 026 → 2 941 lignes.

Contribue à #96.
