+++ voyage-donnees.ts
@@ -1,0 +1,113 @@
+// Le VOYAGE : manifestes de jambe, intentions persistées (ADR-011).
+//
+// Le PARCOURS (les arrêts) va dans l'URL ; les manifestes édités restent LOCAUX. On ne persiste que
+// l'INTENTION de l'utilisateur — `[{ name, units }]` par jambe — jamais un instantané de marché :
+// figé, il continuerait d'afficher le prix du jour de l'édition longtemps après qu'UEX l'ait
+// republié, et la pastille de fraîcheur vieillirait sans refléter le vrai relevé.
+//
+// Ce module porte la DONNÉE — calculer un manifeste, lire une intention, la persister. Les ACTIONS
+// qui la modifient (déplier un éditeur, changer des SCU, ajouter une suggestion) restent dans
+// `app.js` : elles enchaînent une mutation ET un rendu, et c'est le rendu qui les y retient. Elles
+// partiront quand la vue Voyage emménagera dans l'arbre.
+//
+// ── UNE RÉFÉRENCE VIVE SORT D'ICI, ET C'EST VOULU ─────────────────────────────────────────────
+// `legIntent` rend `etat.JOURNEY_EDITS[k]` TEL QUEL, et l'appelant le mute (`intent[li].units = …`).
+// Aucun accesseur, aucun proxy ne verrait cette écriture — c'est l'une des quatre raisons pour
+// lesquelles `etat.ts` ne notifie pas à l'écriture (cf. son en-tête). Ne pas « corriger » ça en
+// rendant une copie : les mutants comptent sur l'identité, et ils appellent `saveJourneyEdits()`
+// eux-mêmes.
+
+import {
+  bestManifest, manifestIntent, manifestIntentSurvives, hydrateManifestLine,
+} from "./logic.ts";
+import type { Filtres, Jambe, LigneManifeste } from "./types.ts";
+import { etat } from "./etat.ts";
+import { stationLabel } from "./logic.ts";
+import { findCommodity, stationMap } from "./marche.ts";
+import { effVals } from "./corrections.ts";
+import { feeCtx, feeResolver } from "./frais.ts";
+
+export function legManifest(leg: Jambe, f: Filtres) {
+  if (!etat.MARKET || !stationMap.size) return null;
+  const fromIdx = stationMap.get(stationLabel(leg.from, leg.fromSystem));
+  const toIdx = stationMap.get(stationLabel(leg.to, leg.toSystem));
+  if (fromIdx == null || toIdx == null) return null;
+  return bestManifest(etat.MARKET, fromIdx, "", f, effVals, toIdx, feeResolver(f)); // { lines, profit, … } ou null
+}
+
+// Contexte de frais d'une jambe. Le récap du voyage est affiché à CÔTÉ des tableaux : le laisser en
+// brut pendant que les six vues passent en net mettrait deux chiffres contradictoires côte à côte.
+export const legFeeCtx = (leg: Jambe, f: Filtres) => feeCtx(f, leg.from, leg.to);
+
+// ---------- Édition inline des manifestes de jambe (persistée en localStorage, HORS lien) ----------
+// Le PARCOURS (arrêts) va dans l'URL ; les manifestes édités restent locaux.
+// On ne persiste que l'INTENTION de l'utilisateur — [{ name, units }] par jambe — jamais un
+// instantané de marché : figé, il continuerait d'afficher le prix du jour de l'édition longtemps
+// après qu'UEX l'ait republié, et la pastille de fraîcheur vieillirait sans refléter le vrai relevé.
+// Prix, stock, demande et dates sont donc RELUS à chaque rendu (cf. hydrateManifestLine).
+// Clé versionnée : l'ancien format stockait des lignes complètes sous une clé « from|to » qui
+// confondait deux jambes identiques d'un même parcours. Les anciennes éditions sont abandonnées.
+const JOURNEY_EDITS_KEY = "best-hauling-journey-edits-v2";
+// Jambes dont les quantités ont été FIGÉES par une correction de volume, et non ajustées à la main.
+// Store séparé plutôt qu'un champ dans JOURNEY_EDITS : le format persisté de l'intention reste
+// intact (aucune migration), et les deux notions se lisent indépendamment. La valeur n'existe que
+// si une entrée d'intention existe au même rang — le gel EST une intention, avec un autre motif.
+const JOURNEY_PINS_KEY = "best-hauling-journey-pins";
+export function loadJourneyPins(): void {
+  try { etat.JOURNEY_PINS = JSON.parse(localStorage.getItem(JOURNEY_PINS_KEY)) || {}; } catch { etat.JOURNEY_PINS = {}; }
+}
+export function saveJourneyPins(): void { try { localStorage.setItem(JOURNEY_PINS_KEY, JSON.stringify(etat.JOURNEY_PINS)); } catch {} }
+export function loadJourneyEdits(): void {
+  try { etat.JOURNEY_EDITS = JSON.parse(localStorage.getItem(JOURNEY_EDITS_KEY)) || {}; } catch { etat.JOURNEY_EDITS = {}; }
+  try { localStorage.removeItem("best-hauling-journey-edits"); } catch {} // format v1 abandonné
+}
+export function saveJourneyEdits(): void { try { localStorage.setItem(JOURNEY_EDITS_KEY, JSON.stringify(etat.JOURNEY_EDITS)); } catch {} }
+// Le RANG de la jambe fait partie de la clé : sans lui, un parcours A→B→A→B partageait un seul
+// manifeste entre ses jambes 1 et 3 (éditer l'une réécrivait l'autre, la supprimer supprimait l'autre).
+export const legKey = (leg: Jambe, i: number): string => `${i}|${leg.from}|${leg.to}`;
+
+// Indices des terminaux d'une jambe, ou null si le marché ne les connaît pas (encore).
+export function legTerminals(leg: Jambe) {
+  const fromIdx = stationMap.get(stationLabel(leg.from, leg.fromSystem));
+  const toIdx = stationMap.get(stationLabel(leg.to, leg.toSystem));
+  return fromIdx == null || toIdx == null ? null : { fromIdx, toIdx };
+}
+
+// Manifeste EFFECTIF d'une jambe : intention éditée ré-hydratée si elle existe, sinon l'optimal.
+export function legEffectiveLines(leg: Jambe, i: number, f: Filtres): LigneManifeste[] {
+  const k = legKey(leg, i);
+  const intent = etat.JOURNEY_EDITS[k];
+  if (!intent) { const man = legManifest(leg, f); return man ? man.lines : []; }
+  const t = legTerminals(leg);
+  if (!etat.MARKET || !t) return [];
+  const lines = [], gardees = [];
+  for (const e of intent) {
+    const c = findCommodity(e.name);
+    if (!c) continue; // commodité disparue d'UEX : on l'oublie plutôt que d'afficher un fantôme
+    gardees.push(e);
+    lines.push(hydrateManifestLine(etat.MARKET, t.fromIdx, t.toIdx, c, e.units, effVals));
+  }
+  // Purge sur place : sans ça l'index des lignes affichées et celui du store divergeraient, et
+  // éditer une quantité écrirait dans la mauvaise entrée.
+  if (gardees.length !== intent.length) { etat.JOURNEY_EDITS[k] = gardees; saveJourneyEdits(); }
+  return lines;
+}
+
+// Bascule la jambe en mode « édité » la 1re fois : on y copie l'intention issue de l'optimal.
+// Toucher au chargement fait de la jambe une édition PERSONNELLE : si elle n'était que figée par
+// une correction de volume, elle cesse de l'être (🔒 -> ✎). Le geste de l'utilisateur prime sur
+// la raison technique qui avait gelé les quantités.
+export function legIntent(leg: Jambe, i: number, f: Filtres) {
+  const k = legKey(leg, i);
+  if (!etat.JOURNEY_EDITS[k]) etat.JOURNEY_EDITS[k] = manifestIntent(legManifest(leg, f)?.lines || []);
+  if (etat.JOURNEY_PINS[k]) { delete etat.JOURNEY_PINS[k]; saveJourneyPins(); }
+  return etat.JOURNEY_EDITS[k];
+}
+
+// Fige les jambes qu'une correction de volume rebattrait, AVANT qu'elle soit appliquée : on capture
+// donc les quantités telles qu'elles sont encore. La sélection est pure (legsToPin) ; ici on ne
+// fournit que ce que logic.mjs ne peut pas connaître — les chargements effectifs du moment.
+// Fige les SCU d'une jambe : son chargement devient une INTENTION persistée, et le 🔒 dit que ce
+// n'est pas la main de l'utilisateur qui l'a voulu. Rend `true` si quelque chose a bougé.
+// Une jambe déjà ajustée (✎) ou déjà figée n'est pas retouchée : ses quantités ne bougeaient plus,
+// et l'écraser effacerait un ajustement fait à la main.
